Support for addition of SUT & Emissary strategies
Split schema and validator from single Job into BrowserApp and Api to
support addition of SUT and Emissary strategies.
Update .env.example to include the env var referencing the new
required source directory.
Update compose file to include new bind mount.
Increase local duration before next Test Run can be initiated from
10 to 15 seconds.
