+++ compose/.env.example
@@ -2,3 +2,7 @@
 # Orchestrator and tester containers share the same group.
 # They also read, write and delete outcomes files within this directory.
 HOST_DIR=</mnt/your-spare-drive/Logs/purpleteam/outcomes>
+# This directory needs group rwx and other r (for the Emissary to read) permissions.
+# Both app-scanner and it's Emissary mount this directory. The app-scanner puts files here for the Emissary to consume.
+# This environment variable and the same value is also required by the App Emissary in the purpleteam-s2-containers project.
+HOST_DIR_APP_SCANNER=</mnt/your-spare-drive/purpleteam-app-scanner>
