+++ src/api/orchestration/models/orchestrate.js
@@ -107,7 +107,6 @@ class Orchestrate {
     return !this.#initTesterResponsesForCli ? 'orchestrator is ready to take orders.' : 'Test Run is in progress.';
   }
 
-  // Could potentially be exposed as part of the API for CLI to invoke if the back-end gets messed up.
   async resetTesters({ level = 'soft' }) {
     const { cleanUpAfterTestRun: cleanUpTesterWatcherAfterTestRun } = this.#testerWatcher;
     const reset = {
