+++ src/api/orchestration/schemas/job.aPi.js
@@ -1,0 +1,9 @@
+const internals = { config: null };
+
+const init = (config) => {
+  internals.config = config;
+};
+
+const schema = {};
+
+module.exports = { init, schema };
